requirements.txt
alabaster==0.7.9
astroid==1.4.8
click==6.6
coverage==4.2
dj-database-url==0.4.1
Django==1.10
django-appconf==1.0.2
django-compressor==2.1
django-debug-toolbar==1.5
django-tables2==1.6.1
djangorestframework==3.6.2
django-filter==1.0.4
docutils==0.12
drfdocs==0.0.11
ecdsa==0.13
Fabric==1.12.0
Faker==0.7.12
gunicorn==19.7.1
first==2.0.1
flake8==3.0.4
imagesize==0.7.1
ipdb==0.9.4
isort==4.2.5
Jinja2==2.8
lazy-object-proxy==1.2.2
MarkupSafe==0.23
mccabe==0.5.2
paramiko==1.17.2
pep8==1.7.0
pip-tools==1.7.0
pycparser==2.14
pycrypto==2.6.1
Pygments==2.1.3
pylint==1.6.4
psycopg2==2.7.1
pytest==3.0.3
pytz==2016.6.1
raven==6.0.0
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
Sphinx==1.4.6
sqlparse==0.2.1
tox==2.4.1
whitenoise==3.2
wrapt==1.10.8
